1 - Death Mode AI changes for vanilla bosses. 2 - Death Mode AI changes for Cal Clone, Ravager, Aquatic Scourge, Brimstone Elemental, Astrum Aureus, Astrum Deus, Ceaseless Void, Dragonfolly and Old Duke.

+++ NPCs/Ravager/RavagerLegLeft.cs
@@ -99,7 +99,7 @@ namespace CalamityMod.NPCs.Ravager
                 float num661 = Main.npc[CalamityGlobalNPC.scavenger].Center.Y - vector79.Y;
                 num661 += 88f;
                 num660 -= 70f;
-                float num662 = (float)Math.Sqrt((double)(num660 * num660 + num661 * num661));
+                float num662 = (float)Math.Sqrt(num660 * num660 + num661 * num661);
                 if (num662 < 12f + num659)
                 {
                     npc.rotation = 0f;
